requirements.txt
autopep8==1.5.7
bandit==1.7.0
gitdb==4.0.7
GitPython==3.1.24
pbr==5.6.0
pycodestyle==2.8.0
python-dotenv==0.19.1
PyYAML==6.0
six==1.16.0
smmap==4.0.0
stevedore==3.4.0
toml==0.10.2
typing-extensions==3.10.0.2
